streamlit/requirements.txt
image-similarity-measures == 0.3.6
numpy == 1.24.4
opencv-python == 4.8.0.74
pandas == 2.0.3
pyfftw == 0.13.1
pymongo == 4.4.*
python-dotenv == 1.0.0
streamlit == 1.24.0